requirements.txt
h5py==3.10.*
nflows @ git+https://github.com/ThorstenBuss/nflows.git@138f86830e161bb0095dbb5cdd51871b769ea5b7
numpy==1.26.*
matplotlib==3.8.*
PyYAML==6.0.*
scikit-learn==1.3.*
scipy==1.11.*
torch==2.0.*
torchinfo==1.8.*
torchmetrics==1.2.*
